requirements.txt
wandb
pycocotools 
ftfy 
regex 
tqdm 
ipywidgets
seaborn
einops
clip-anytorch
pycocotools
kornia
